added ability to sign up and login

+++ Backend/controllers/todo.controller.js
@@ -3,8 +3,9 @@ import { Todo } from '../models/todo.model.js'
 export const createTodo = async (req, res, next) => {
     try{
           const { description } = req.body
+          const userId = req.user._id
 
-          const todoItem = await Todo.create({description })
+          const todoItem = await Todo.create({description, userId })
 
           res.status(200).json({
                success:'added successfully',
@@ -21,7 +22,8 @@ export const createTodo = async (req, res, next) => {
 
 export const getTodos = async (req, res, next) => {
      try{
-          const todoItem = await Todo.find()
+          const userId = req.user._id
+          const todoItem = await Todo.find({userId})
 
           res.status(200).json({
                success:'all todos gotten',
@@ -37,15 +39,16 @@ export const getTodos = async (req, res, next) => {
 
 export const deleteTodo = async (req, res, next) => {
      try{
-          const todoItem = await Todo.findByIdAndDelete(req.params.id)
+          const userId = req.user._id
+          const todoItem = await Todo.findOneAndDelete({_id: req.params.id, userId})
           
           if(!todoItem){
-               return res.status(404).json({success:false, message:"Todo not found"})
+               return res.status(404).json({success:false, message:"Todo not found or not owned by user"})
           }
 
            res.status(200).json({
                success: true,
-               message: 'Todo deleted successfully' 
+               message: 'Todo deleted successfully'
           });
      }
      catch(err){
@@ -56,11 +59,12 @@ export const deleteTodo = async (req, res, next) => {
 
 export const deleteTodos = async (req, res, next) => {
      try{
-               await Todo.deleteMany({})
+          const userId = req.user._id
+               await Todo.deleteMany({userId})
 
            res.status(200).json({
                success: true,
-               message: 'All todos deleted successfully' 
+               message: 'All todos deleted successfully'
           });
      }
      catch(err){
